Optimized Db Lollipop

+++ app/src/main/java/com/smartsolutions/paquetes/workers/TrafficDbOptimizerLollipop.kt
@@ -1,8 +1,11 @@
 package com.smartsolutions.paquetes.workers
 
 import android.content.Context
+import android.os.Build
 import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
+import androidx.work.PeriodicWorkRequestBuilder
+import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import com.smartsolutions.paquetes.helpers.DateCalendarUtils
 import com.smartsolutions.paquetes.managers.contracts.ISimManager
@@ -10,6 +13,7 @@ import com.smartsolutions.paquetes.managers.models.Traffic
 import com.smartsolutions.paquetes.repositories.contracts.ITrafficRepository
 import dagger.assisted.Assisted
 import org.apache.commons.lang.time.DateUtils
+import java.util.concurrent.TimeUnit
 
 @HiltWorker
 class TrafficDbOptimizerLollipop(
@@ -18,42 +22,64 @@ class TrafficDbOptimizerLollipop(
     @Assisted
     parameters: WorkerParameters,
     private val trafficRepository: ITrafficRepository,
-    private val simManager: ISimManager,
-    private val dateCalendarUtils: DateCalendarUtils
+    private val simManager: ISimManager
 ): CoroutineWorker(context, parameters) {
 
 
     override suspend fun doWork(): Result {
 
         val traffics = mutableListOf<Traffic>()
-        val dayPeriod = dateCalendarUtils.getTimePeriod(DateCalendarUtils.PERIOD_TODAY)
 
         simManager.getInstalledSims().forEach { sim ->
             var traffic: Traffic? = null
-            var periodFinish = 0L
+            var hour = DateCalendarUtils.getStartAndFinishHour(System.currentTimeMillis())
 
-            val oldTraffics = trafficRepository.getByTime(sim.id, dayPeriod.first, dayPeriod.second).onEach { trafficDb ->
+            val oldTraffics = trafficRepository.getAll(sim.id).onEach { trafficDb ->
                 when {
                     traffic == null -> {
                         traffic = trafficDb
-                        periodFinish = trafficDb.startTime + DateUtils.MILLIS_PER_HOUR
+                        hour = DateCalendarUtils.getStartAndFinishHour(trafficDb.startTime)
                     }
-                    trafficDb.startTime <= periodFinish -> {
+                    trafficDb.startTime in hour.first..hour.second -> {
                         traffic!! += trafficDb
                     }
                     else -> {
                         traffics.add(traffic!!)
                         traffic = trafficDb
-                        periodFinish = trafficDb.startTime + DateUtils.MILLIS_PER_HOUR
+                        hour = DateCalendarUtils.getStartAndFinishHour(trafficDb.startTime)
                     }
                 }
             }
 
-            trafficRepository.delete(oldTraffics)
-            trafficRepository.create(traffics)
-            traffics.clear()
+            if (oldTraffics.isNotEmpty() && traffics.isNotEmpty()) {
+                trafficRepository.delete(oldTraffics)
+                trafficRepository.create(traffics)
+                traffics.clear()
+            }
         }
 
         return Result.success()
+    }
+
+
+    companion object {
+
+        private const val TAG_OPTIMIZER_WORKER = "tag_optimizer_worker"
+
+        fun registerWorkerIfNeeded(context: Context, hoursInterval: Long){
+            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M){
+                val request = PeriodicWorkRequestBuilder<TrafficDbOptimizerLollipop>(
+                    hoursInterval,
+                    TimeUnit.HOURS
+                )
+                    .addTag(TAG_OPTIMIZER_WORKER)
+                    .build()
+
+                val workManager = WorkManager.getInstance(context)
+                workManager.cancelAllWorkByTag(TAG_OPTIMIZER_WORKER)
+                workManager.enqueue(request)
+            }
+        }
+
     }
 }
